+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/define_fact/define_fact_test.go
@@ -27,12 +27,15 @@ func TestDefineFactInstruction_GetCanonicalizedInstruction(t *testing.T) {
 		testFactName,
 		emptyFactsRecipe,
 	)
-	expectedFormatStr := `# from: dummyFile[1:1]
+	expectedMultiLineFormatStr := `# from: dummyFile[1:1]
 define_fact(
 	fact_name="%v",
-	service_id="%v",
+	service_id="%v"
 )`
-	expectedStr := fmt.Sprintf(expectedFormatStr, testFactName, testServiceId)
-	require.Equal(t, expectedStr, execInstruction.GetCanonicalInstruction())
-	require.Equal(t, expectedStr, execInstruction.String())
+	expectedMultiLineStr := fmt.Sprintf(expectedMultiLineFormatStr, testFactName, testServiceId)
+	require.Equal(t, expectedMultiLineStr, execInstruction.GetCanonicalInstruction())
+
+	expectedSingleLineFormatStr := `define_fact(fact_name="%v", service_id="%v")`
+	expectedSingleLineStr := fmt.Sprintf(expectedSingleLineFormatStr, testFactName, testServiceId)
+	require.Equal(t, expectedSingleLineStr, execInstruction.String())
 }