+++ internal_testsuites/golang/testsuite/stream_logs_test/stream_logs_test.go
@@ -8,6 +8,7 @@ package stream_logs_test
 import (
 	"context"
 	"github.com/kurtosis-tech/kurtosis-cli/golang_internal_testsuite/test_helpers"
+	"github.com/kurtosis-tech/kurtosis/api/golang/core/lib/enclaves"
 	"github.com/kurtosis-tech/kurtosis/api/golang/core/lib/services"
 	"github.com/kurtosis-tech/kurtosis/api/golang/engine/lib/kurtosis_context"
 	"github.com/kurtosis-tech/stacktrace"
@@ -28,17 +29,22 @@ const (
 
 	testTimeOut = 90 * time.Second
 
-	shouldFollowLogsFirstRequest  = false
-	shouldFollowLogsSecondRequest = true
+	shouldFollowLogs  = true
+	shouldNotFollowLogs = false
+
+	nonExistentServiceGuid = "stream-logs-1667939326-non-existent"
 )
 
-var	shouldFollowLogsRequestOptions = []bool{shouldFollowLogsFirstRequest, shouldFollowLogsSecondRequest}
+type serviceLogsRequestInfoAndExpectedResults struct {
+	requestedEnclaveID    enclaves.EnclaveID
+	requestedServiceGuids map[services.ServiceGUID]bool
+	requestedFollowLogs   bool
+	expectedLogLines []string
+	expectedNotFoundServiceGuids map[services.ServiceGUID]bool
+}
 
 func TestStreamLogs(t *testing.T) {
 	ctx := context.Background()
-
-	expectedLogLines := []string{"kurtosis", "test", "running", "successfully"}
-	expectedAmountOfLogLines := len(expectedLogLines)
 
 	// ------------------------------------- ENGINE SETUP ----------------------------------------------
 	enclaveCtx, stopEnclaveFunc, _, err := test_helpers.CreateEnclave(t, ctx, testName, isPartitioningEnabled)
@@ -60,49 +66,69 @@ func TestStreamLogs(t *testing.T) {
 	serviceCtx, err := enclaveCtx.GetServiceContext(exampleServiceId)
 	require.NoError(t, err)
 
-	userServiceGuid := serviceCtx.GetServiceGUID()
+	serviceGuid := serviceCtx.GetServiceGUID()
 
-	userServiceGuids := map[services.ServiceGUID]bool{
-		userServiceGuid: true,
+	serviceGuids := map[services.ServiceGUID]bool{
+		serviceGuid: true,
 	}
 
 	time.Sleep(waitForAllLogsBeingCollectedInSeconds * time.Second)
 
-	// will test executing getUserServiceLogs with shouldFollowLogs = false in the first iteration,
-	// and with shouldFollowLogs = true (which is used to tail logs) in the second iteration
-	for _, shouldFollowLogs := range shouldFollowLogsRequestOptions {
-		//TODO handle notFOundChannel
-		userServiceLogsByGuidChan, cancelStreamUserServiceLogsFunc, err := kurtosisCtx.GetUserServiceLogs(ctx, enclaveID, userServiceGuids, shouldFollowLogs)
+	serviceLogsRequestInfoAndExpectedResultsList := getServiceLogsRequestInfoAndExpectedResultsList(
+		enclaveID,
+		serviceGuids,
+	)
+
+	for _, serviceLogsRequestInfoAndExpectedResultsObj := range serviceLogsRequestInfoAndExpectedResultsList {
+
+		requestedEnclaveId := serviceLogsRequestInfoAndExpectedResultsObj.requestedEnclaveID
+		requestedServiceGuids := serviceLogsRequestInfoAndExpectedResultsObj.requestedServiceGuids
+		requestedShouldFollowLogs := serviceLogsRequestInfoAndExpectedResultsObj.requestedFollowLogs
+		expectedLogLines := serviceLogsRequestInfoAndExpectedResultsObj.expectedLogLines
+		expectedNonExistenceServiceGuids := serviceLogsRequestInfoAndExpectedResultsObj.expectedNotFoundServiceGuids
+
+		serviceLogsStreamContentChan, cancelStreamServiceLogsFunc, err := kurtosisCtx.GetServiceLogs(ctx, requestedEnclaveId, requestedServiceGuids, requestedShouldFollowLogs)
 		require.NoError(t, err)
-		require.NotNil(t, cancelStreamUserServiceLogsFunc)
-		require.NotNil(t, userServiceLogsByGuidChan)
+		require.NotNil(t, cancelStreamServiceLogsFunc)
+		require.NotNil(t, serviceLogsStreamContentChan)
 
 		receivedLogLines := []string{}
+		receivedNotFoundServiceGuids := map[services.ServiceGUID]bool{}
 
 		var testEvaluationErr error
 
-		shouldReceiveStream := true
-		for shouldReceiveStream {
+		shouldContinueInTheLoop := true
+
+		for shouldContinueInTheLoop {
 			select {
 			case <-time.Tick(testTimeOut):
 				testEvaluationErr = stacktrace.NewError("Receiving stream logs in the test has reached the '%v' time out", testTimeOut)
-				shouldReceiveStream = false
+				shouldContinueInTheLoop = false
 				break
-			case userServiceLogsByGuid, isChanOpen := <-userServiceLogsByGuidChan:
+			case serviceLogsStreamContent, isChanOpen := <-serviceLogsStreamContentChan:
 				if !isChanOpen {
-					shouldReceiveStream = false
+					shouldContinueInTheLoop = false
 					break
 				}
 
-				userServiceLogs, found := userServiceLogsByGuid[userServiceGuid]
-				require.True(t, found)
+				serviceLogsByGuid := serviceLogsStreamContent.GetServiceLogsByServiceGuids()
+				notFoundGuids := serviceLogsStreamContent.GetNotFoundServiceGuids()
 
-				for _, serviceLog := range userServiceLogs {
+				serviceLogLines, found := serviceLogsByGuid[serviceGuid]
+				if len(expectedNonExistenceServiceGuids) > 0 {
+					require.False(t, found)
+				} else {
+					require.True(t, found)
+				}
+
+				receivedNotFoundServiceGuids = notFoundGuids
+
+				for _, serviceLog := range serviceLogLines {
 					receivedLogLines = append(receivedLogLines, serviceLog.GetContent())
 				}
 
-				if len(receivedLogLines) == expectedAmountOfLogLines {
-					shouldReceiveStream = false
+				if len(receivedLogLines) == len(expectedLogLines) {
+					shouldContinueInTheLoop = false
 					break
 				}
 			}
@@ -110,7 +136,8 @@ func TestStreamLogs(t *testing.T) {
 
 		require.NoError(t, testEvaluationErr)
 		require.Equal(t, expectedLogLines, receivedLogLines)
-		cancelStreamUserServiceLogsFunc()
+		require.Equal(t, expectedNonExistenceServiceGuids, receivedNotFoundServiceGuids)
+		cancelStreamServiceLogsFunc()
 	}
 }
 
@@ -131,4 +158,52 @@ func getExampleServiceConfig() *services.ContainerConfig {
 		cmdArgs,
 	).Build()
 	return containerConfig
+}
+
+func getServiceLogsRequestInfoAndExpectedResultsList(
+	enclaveId enclaves.EnclaveID,
+	serviceGuids map[services.ServiceGUID]bool,
+) []*serviceLogsRequestInfoAndExpectedResults {
+
+	expectedLogLineValues := []string{"kurtosis", "test", "running", "successfully"}
+	expectedEmptyLogLineValues := []string{}
+
+	emptyServiceGuids := map[services.ServiceGUID]bool{}
+
+	nonExistentServiceGuids := map[services.ServiceGUID]bool{
+		nonExistentServiceGuid: true,
+	}
+
+	firstCallRequestInfoAndExpectedResults := &serviceLogsRequestInfoAndExpectedResults{
+		requestedEnclaveID:           enclaveId,
+		requestedServiceGuids:        serviceGuids,
+		requestedFollowLogs:          shouldNotFollowLogs,
+		expectedLogLines:             expectedLogLineValues,
+		expectedNotFoundServiceGuids: emptyServiceGuids,
+
+	}
+
+	secondCallRequestInfoAndExpectedResults := &serviceLogsRequestInfoAndExpectedResults{
+		requestedEnclaveID:           enclaveId,
+		requestedServiceGuids:        serviceGuids,
+		requestedFollowLogs:          shouldFollowLogs,
+		expectedLogLines:             expectedLogLineValues,
+		expectedNotFoundServiceGuids: emptyServiceGuids,
+	}
+
+	thirdCallRequestInfoAndExpectedResults := &serviceLogsRequestInfoAndExpectedResults{
+		requestedEnclaveID:           enclaveId,
+		requestedServiceGuids:        nonExistentServiceGuids,
+		requestedFollowLogs:          shouldFollowLogs,
+		expectedLogLines:             expectedEmptyLogLineValues,
+		expectedNotFoundServiceGuids: nonExistentServiceGuids,
+	}
+
+	serviceLogsRequestInfoAndExpectedResultsList := []*serviceLogsRequestInfoAndExpectedResults{
+		firstCallRequestInfoAndExpectedResults,
+		secondCallRequestInfoAndExpectedResults,
+		thirdCallRequestInfoAndExpectedResults,
+	}
+
+	return serviceLogsRequestInfoAndExpectedResultsList
 }